Hacked circular dependency, give Lambda upload rights to create URL
We're building the ARNs for predictable resources, but some like
StateMachine are generated with unique suffixes so we can't build
those. This solves the circular dependency at least.

We may be able to get around this by (say) defining the Bucket,
Lambdas, and any Policies, and using "depends_on" in the SM so it
will build those first.

I also realized why I (and Hong) were getting access denied with the
PSURL: the Lambda has to be given rights to do the put_object on the
bucket, else the PSURL won't be honored. Now we can upload, duh.