Added error handling states to update DDB, S3, then fail.
I expect we can have all error handlers set a fail to the statemachine
and go to a single Failed handler that captures their output so we can
skip the specific *Fail states.

I'm still not seeing how to handle timeouts waiting for the upload,
and the docs aren't clear. It sounds like Timeout is used to wait for
our 'Resource' -- a lambda -- to complete. But we really want the
machine to STOP with nothing running, and only begin again once the
upload comes in; if it doesn't come in within an hour, we go to a
Timeout error state and cleanup.